requirements.txt
fastapi>=0.100.0
uvicorn>=0.23.0
python-multipart>=0.0.6
pydantic>=2.0.0
python-dotenv>=1.0.0
openpyxl>=3.1.0
pandas>=2.0.0
aiofiles>=23.0.0
passlib>=1.7.4
bcrypt>=4.0.0
PyJWT>=2.8.0
httpx>=0.25.0
requests>=2.31.0
opencv-python>=4.8.0
numpy>=1.24.0